+++ packages/terra-notification-dialog/src/NotificationDialog.jsx
@@ -5,6 +5,7 @@ import FocusTrap from 'focus-trap-react';
 import KeyCode from 'keycode-js';
 import Button from 'terra-button';
 import classNames from 'classnames/bind';
+import { FormattedMessage } from 'react-intl';
 import styles from './NotificationDialog.module.scss';
 
 const cx = classNames.bind(styles);
@@ -94,7 +95,7 @@ const actionSection = (primaryAction, secondaryAction) => {
   );
 };
 
-const getIcon = (intl, variant, customIcon = null) => {
+const getIcon = (variant, customIcon = null) => {
   switch (variant) {
     case variants.ALERT:
       return (<span className={cx(['icon', 'alert'])} />);
@@ -111,15 +112,6 @@ const getIcon = (intl, variant, customIcon = null) => {
     default:
       return null;
   }
-};
-
-const contextTypes = {
-  /* eslint-disable consistent-return */
-  intl: (context) => {
-    if (context.intl === undefined) {
-      return new Error('Please add locale prop to Base component to load translations');
-    }
-  },
 };
 
 class NotificationDialog extends React.Component {
@@ -167,9 +159,7 @@ class NotificationDialog extends React.Component {
       ...customProps
     } = this.props;
 
-    const { intl } = this.context;
-
-    const defaultHeader = variant === variants.CUSTOM ? '' : intl.formatMessage({ id: `Terra.notification.dialog.${variant}` });
+    const defaultHeader = variant === variants.CUSTOM ? '' : <FormattedMessage id={`Terra.notification.dialog.${variant}`} />;
     const notificationDialogClassNames = cx('notification-dialog', customProps.className);
 
     return (
@@ -192,15 +182,12 @@ class NotificationDialog extends React.Component {
               <div id="notification-dialog-header" className={cx('header-body')}>{header || defaultHeader}</div>
               <div className={cx('notification-dialog-body')}>
                 {variant
-                  && <div className={cx('icon-container')}>{getIcon(intl, variant, customIcon)}</div>
-                }
+                  && <div className={cx('icon-container')}>{getIcon(variant, customIcon)}</div>}
                 <div className={cx('text-wrapper')}>
                   {title
-                    && <div id="notification-dialog-title" className={cx('title')}>{title}</div>
-                  }
+                    && <div id="notification-dialog-title" className={cx('title')}>{title}</div>}
                   {message
-                    && <div className={cx('message')}>{message}</div>
-                  }
+                    && <div className={cx('message')}>{message}</div>}
                 </div>
               </div>
               <div className={cx('footer-body')}>{actionSection(primaryAction, secondaryAction)}</div>
@@ -214,7 +201,6 @@ class NotificationDialog extends React.Component {
 
 NotificationDialog.propTypes = propTypes;
 NotificationDialog.defaultProps = defaultProps;
-NotificationDialog.contextTypes = contextTypes;
 
 export { variants as NotificationDialogVariants };
 export default NotificationDialog;