+++ packages/terra-time-input/src/TimeInput.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import Input from 'terra-form-input';
 import ButtonGroup from 'terra-button-group';
+import { injectIntl, intlShape } from 'react-intl';
 
 import KeyCode from 'keycode-js';
 import TimeUtil from './TimeUtil';
@@ -25,6 +26,11 @@ const propTypes = {
    */
   // eslint-disable-next-line react/forbid-prop-types
   hourAttributes: PropTypes.object,
+  /**
+  * @private
+  * intl object programmatically imported through injectIntl from react-intl.
+  * */
+  intl: intlShape.isRequired,
   /**
    * Custom input attributes to apply to the minutes input
    */
@@ -87,17 +93,8 @@ const defaultProps = {
   variant: TimeUtil.FORMAT_24_HOUR,
 };
 
-const contextTypes = {
-  /* eslint-disable consistent-return */
-  intl: (context) => {
-    if (context.intl === undefined) {
-      return new Error('Component is internationalized, and must be wrapped in terra-base');
-    }
-  },
-};
-
 class TimeInput extends React.Component {
-  constructor(props, context) {
+  constructor(props) {
     super(props);
 
     let { value } = this.props;
@@ -141,7 +138,7 @@ class TimeInput extends React.Component {
     let meridiem;
 
     if (props.variant === TimeUtil.FORMAT_12_HOUR) {
-      if (!context.intl.messages['Terra.timeInput.am'] || !context.intl.messages['Terra.timeInput.pm']) {
+      if (!this.props.intl.messages['Terra.timeInput.am'] || !this.props.intl.messages['Terra.timeInput.pm']) {
         if (process.env !== 'production') {
           // eslint-disable-next-line no-console
           console.warn('This locale only uses 24 hour clock. The ante meridiem and post meridiem will not be displayed');
@@ -150,8 +147,8 @@ class TimeInput extends React.Component {
         this.anteMeridiem = '';
         this.postMeridiem = '';
       } else {
-        this.anteMeridiem = context.intl.formatMessage({ id: 'Terra.timeInput.am' });
-        this.postMeridiem = context.intl.formatMessage({ id: 'Terra.timeInput.pm' });
+        this.anteMeridiem = this.props.intl.formatMessage({ id: 'Terra.timeInput.am' });
+        this.postMeridiem = this.props.intl.formatMessage({ id: 'Terra.timeInput.pm' });
       }
 
       if (hour) {
@@ -189,7 +186,7 @@ class TimeInput extends React.Component {
     let { meridiem } = this.state;
 
     if (this.props.variant === TimeUtil.FORMAT_12_HOUR) {
-      if (!this.context.intl.messages['Terra.timeInput.am'] || !this.context.intl.messages['Terra.timeInput.pm']) {
+      if (!this.props.intl.messages['Terra.timeInput.am'] || !this.props.intl.messages['Terra.timeInput.pm']) {
         if (process.env !== 'production') {
           // eslint-disable-next-line no-console
           console.warn('This locale only uses 24 hour clock. The ante meridiem and post meridiem will not be displayed');
@@ -198,8 +195,8 @@ class TimeInput extends React.Component {
         this.anteMeridiem = '';
         this.postMeridiem = '';
       } else {
-        this.anteMeridiem = this.context.intl.formatMessage({ id: 'Terra.timeInput.am' });
-        this.postMeridiem = this.context.intl.formatMessage({ id: 'Terra.timeInput.pm' });
+        this.anteMeridiem = this.props.intl.formatMessage({ id: 'Terra.timeInput.am' });
+        this.postMeridiem = this.props.intl.formatMessage({ id: 'Terra.timeInput.pm' });
       }
 
       if (hour) {
@@ -709,6 +706,7 @@ class TimeInput extends React.Component {
       inputAttributes,
       minuteAttributes,
       hourAttributes,
+      intl,
       onBlur,
       onChange,
       onFocus,
@@ -721,9 +719,9 @@ class TimeInput extends React.Component {
       ...customProps
     } = this.props;
 
-    const instanceHoursAttrs = Object.assign({}, hourAttributes);
-    const instanceMinuteAttrs = Object.assign({}, minuteAttributes);
-    const instanceSecondAttrs = Object.assign({}, secondAttributes);
+    const instanceHoursAttrs = { ...hourAttributes };
+    const instanceMinuteAttrs = { ...minuteAttributes };
+    const instanceSecondAttrs = { ...secondAttributes };
 
     // Using the state of hour, minute, and second (if shown) create a time in UTC represented in ISO 8601 format.
     let timeValue = '';
@@ -774,7 +772,7 @@ class TimeInput extends React.Component {
             className={cx('time-input-hour')}
             value={this.state.hour}
             name={'terra-time-hour-'.concat(name)}
-            placeholder={this.context.intl.formatMessage({ id: 'Terra.timeInput.hh' })}
+            placeholder={intl.formatMessage({ id: 'Terra.timeInput.hh' })}
             maxLength="2"
             onChange={this.handleHourChange}
             onKeyDown={this.handleHourInputKeyDown}
@@ -785,7 +783,7 @@ class TimeInput extends React.Component {
             disabled={disabled}
           />
           <label htmlFor={instanceHoursAttrs.id} className={cx('mobile-input-label')}>
-            {this.context.intl.formatMessage({ id: 'Terra.timeInput.hours' })}
+            {intl.formatMessage({ id: 'Terra.timeInput.hours' })}
           </label>
         </div>
         <span className={cx('time-spacer')}>:</span>
@@ -797,7 +795,7 @@ class TimeInput extends React.Component {
             className={cx('time-input-minute', showSeconds ? 'with-second' : 'without-second')}
             value={this.state.minute}
             name={'terra-time-minute-'.concat(name)}
-            placeholder={this.context.intl.formatMessage({ id: 'Terra.timeInput.mm' })}
+            placeholder={intl.formatMessage({ id: 'Terra.timeInput.mm' })}
             maxLength="2"
             onChange={this.handleMinuteChange}
             onKeyDown={this.handleMinuteInputKeyDown}
@@ -808,7 +806,7 @@ class TimeInput extends React.Component {
             disabled={disabled}
           />
           <label htmlFor={instanceMinuteAttrs.id} className={cx('mobile-input-label')}>
-            {this.context.intl.formatMessage({ id: 'Terra.timeInput.minutes' })}
+            {intl.formatMessage({ id: 'Terra.timeInput.minutes' })}
           </label>
         </div>
         {showSeconds && (
@@ -822,7 +820,7 @@ class TimeInput extends React.Component {
                 className={cx('time-input-second')}
                 value={this.state.second}
                 name={'terra-time-second-'.concat(name)}
-                placeholder={this.context.intl.formatMessage({ id: 'Terra.timeInput.ss' })}
+                placeholder={intl.formatMessage({ id: 'Terra.timeInput.ss' })}
                 maxLength="2"
                 onChange={this.handleSecondChange}
                 onKeyDown={this.handleSecondInputKeyDown}
@@ -833,7 +831,7 @@ class TimeInput extends React.Component {
                 disabled={disabled}
               />
               <label htmlFor={instanceSecondAttrs.id} className={cx('mobile-input-label')}>
-                {this.context.intl.formatMessage({ id: 'Terra.timeInput.seconds' })}
+                {intl.formatMessage({ id: 'Terra.timeInput.seconds' })}
               </label>
             </div>
           </React.Fragment>
@@ -872,6 +870,7 @@ class TimeInput extends React.Component {
       inputAttributes,
       minuteAttributes,
       hourAttributes,
+      intl,
       onBlur,
       onChange,
       onFocus,
@@ -925,7 +924,7 @@ class TimeInput extends React.Component {
         <Input
           {...inputAttributes}
           {...minuteAttributes}
-          aria-label={this.context.intl.formatMessage({ id: 'Terra.timeInput.hours' })}
+          aria-label={intl.formatMessage({ id: 'Terra.timeInput.hours' })}
           refCallback={(inputRef) => {
             this.hourInput = inputRef;
             if (refCallback) refCallback(inputRef);
@@ -934,7 +933,7 @@ class TimeInput extends React.Component {
           type="text"
           value={this.state.hour}
           name={'terra-time-hour-'.concat(name)}
-          placeholder={this.context.intl.formatMessage({ id: 'Terra.timeInput.hh' })}
+          placeholder={intl.formatMessage({ id: 'Terra.timeInput.hh' })}
           maxLength="2"
           onChange={this.handleHourChange}
           onKeyDown={this.handleHourInputKeyDown}
@@ -949,12 +948,12 @@ class TimeInput extends React.Component {
           {...inputAttributes}
           {...minuteAttributes}
           refCallback={(inputRef) => { this.minuteInput = inputRef; }}
-          aria-label={this.context.intl.formatMessage({ id: 'Terra.timeInput.minutes' })}
+          aria-label={intl.formatMessage({ id: 'Terra.timeInput.minutes' })}
           className={cx('time-input-minute', showSeconds ? 'with-second' : 'without-second', 'desktop', { 'initial-focus': this.state.minuteInitialFocused })}
           type="text"
           value={this.state.minute}
           name={'terra-time-minute-'.concat(name)}
-          placeholder={this.context.intl.formatMessage({ id: 'Terra.timeInput.mm' })}
+          placeholder={intl.formatMessage({ id: 'Terra.timeInput.mm' })}
           maxLength="2"
           onChange={this.handleMinuteChange}
           onKeyDown={this.handleMinuteInputKeyDown}
@@ -971,12 +970,12 @@ class TimeInput extends React.Component {
               {...inputAttributes}
               {...secondAttributes}
               refCallback={(inputRef) => { this.secondInput = inputRef; }}
-              aria-label={this.context.intl.formatMessage({ id: 'Terra.timeInput.seconds' })}
+              aria-label={intl.formatMessage({ id: 'Terra.timeInput.seconds' })}
               className={cx('time-input-second', 'desktop', { 'initial-focus': this.state.secondInitialFocused })}
               type="text"
               value={this.state.second}
               name={'terra-time-second-'.concat(name)}
-              placeholder={this.context.intl.formatMessage({ id: 'Terra.timeInput.ss' })}
+              placeholder={intl.formatMessage({ id: 'Terra.timeInput.ss' })}
               maxLength="2"
               onChange={this.handleSecondChange}
               onKeyDown={this.handleSecondInputKeyDown}
@@ -992,7 +991,7 @@ class TimeInput extends React.Component {
           [
             <Input
               {...inputAttributes}
-              aria-label={this.context.intl.formatMessage({ id: 'Terra.timeInput.display.meridiem' })} // value in translations set to 'Display Meridiem'
+              aria-label={intl.formatMessage({ id: 'Terra.timeInput.display.meridiem' })} // value in translations set to 'Display Meridiem'
               aria-readonly
               className={cx(['meridiem-display', { focused: this.state.meridiemFocused }])}
               onFocus={this.handleMeridiemInputFocus}
@@ -1008,7 +1007,7 @@ class TimeInput extends React.Component {
               key="meridiem_select_box"
             >
               <select
-                aria-label={this.context.intl.formatMessage({ id: 'Terra.timeInput.select.meridiem' })} // value in translations set to 'Select Meridiem'
+                aria-label={intl.formatMessage({ id: 'Terra.timeInput.select.meridiem' })} // value in translations set to 'Select Meridiem'
                 ref={(select) => { this.meridiemSelect = select; }}
                 onBlur={this.handleMeridiemBlur}
                 onFocus={this.handleMeridiemSelectFocus}
@@ -1041,6 +1040,5 @@ class TimeInput extends React.Component {
 
 TimeInput.propTypes = propTypes;
 TimeInput.defaultProps = defaultProps;
-TimeInput.contextTypes = contextTypes;
 
-export default TimeInput;
+export default injectIntl(TimeInput);