+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -31,10 +31,11 @@ Unreleased
 * Added documentation on required peerDependencies
 
 ### Fixed
+* Fixed divider transition between normal and in-menu view.
 * Fixed Collapsible issue for single item group
 
 6.5.0 - (July 10, 2019)
-------------------
+------------------ 
 ### Changed
 * Minor dependency version bump
 * Replaced Terra-Form with Terra-Form-Checkbox in CollapsibleMenuViewToggle