+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Replaced Object.assign syntax with Object spread syntax
 
 6.9.0 - (August 14, 2019)
 ------------------
@@ -35,7 +37,7 @@ Unreleased
 * Fixed Collapsible issue for single item group
 
 6.5.0 - (July 10, 2019)
------------------- 
+------------------
 ### Changed
 * Minor dependency version bump
 * Replaced Terra-Form with Terra-Form-Checkbox in CollapsibleMenuViewToggle