+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -3,8 +3,17 @@ ChangeLog
 
 Unreleased
 ----------
+
+6.10.0 - (August 21, 2019)
+------------------
+### Changed
+* Replaced Object.assign syntax with Object spread syntax
+
+6.9.0 - (August 14, 2019)
+------------------
 ### Changed
 * updated package.json test scripts
+* Updated to interface with react-intl's `FormattedMessage` rather than old React Context API.
 
 6.8.0 - (July 30, 2019)
 ------------------
@@ -27,6 +36,7 @@ Unreleased
 * Added documentation on required peerDependencies
 
 ### Fixed
+* Fixed divider transition between normal and in-menu view.
 * Fixed Collapsible issue for single item group
 
 6.5.0 - (July 10, 2019)